Moved all components to a component folder, fixed imports
