+++ index.html
@@ -11,7 +11,7 @@
 
 <body>
   <div id="root"></div>
-  <script type="module" src="/src/main.jsx"></script>
+  <script type="module" src="/src/components/main.jsx"></script>
 </body>
 
 </html>
