+++ src/components/Footer.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import './css/style.css';
 
 export default function Footer() {
