+++ src/components/MovieDetail.jsx
@@ -1,4 +1,4 @@
-import { useStates } from './utilities/states';
+import { useStates } from '../utilities/states';
 import { useParams } from 'react-router-dom';
 import  React from 'react';
