+++ src/components/main.jsx
@@ -1,5 +1,5 @@
-import './css/style.css';
-import './utilities/auto-key-lists';
+import '../css/style.css';
+import '../utilities/auto-key-lists';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
